Default recipe, configure rundeck with mysql backing
If the attribute rundeck.use_mysql is set to true(default)
the cookbook configures rundeck to use mysql as a database
instead of using files on the filesystem. If the connection
address of mongo is localhost the recipe goes ahead and
install mysql, adds the rundeck user and the rundeck database
and give rundeck user access to rundeck database. A lot
of rundeck in that last sentence..